+++ hareware/pwm.h
@@ -6,7 +6,7 @@
 
 void PWM_Init(void);
 
-void PWMA_SetDutyCycle(uint8_t dutyCycle);
+void PWMA_SetDutyCycle(int dutyCycle);
 
-void PWMB_SetDutyCycle(uint8_t dutyCycle);
+void PWMB_SetDutyCycle(int dutyCycle);
 #endif // !PWM_H
